services/auth/requirements.txt
aerich
fastapi
PyJWT
uvicorn
redis
redis[hiredis]   
pytest
tortoise-orm
tortoise-orm[asyncpg]
python-dotenv
passlib